Working on changing articles
The functionality to change articles has now been implemented, I decided to revise my getArticleJSON method to return a promise, which makes it easier to handle the asynchronous http request. I have also added a timestamp to the end of the picsum link so each image slide is a new image instead of the same one. Finally I added the list section to the article as well as the correct style for titles and headings