A large portion of task is now complete.
I have added the ability to view all articles as well as move to and from the ratings page which can be viewed at the end of the articles. The ratings are sent though a http post sub. I have commented out the code I would have used if i was able to make a post request to a server. I have begun testing using QUnit. I have also added headings to the readme which i will fill in at the end of this task.