I have mostly worked on styling and the readme section.
I decided i wanted to elaborate on my experience with this problem so i have written a few sections within the readme of this repository. I have also decided i will not have time to implement the tests in the way i would like to do them so i will write a few unit tests to test the main functions of the javascript. I will also write the code i would use as well as how i could carry out preloading the articles as i do not believe i have enough time to implement it.