Working on preloading articles
Surrounding articles are now preloaded and added to array, this way the user does not need to wait for the next article to be loaded. The only time a user is able to view an unloaded article is when first loading the window. I will do some more testing and refactoring before merging branch.